updated home work with all bug fixed and more DRY code
